requirements.txt
ansible==5.8.0
ansible-compat==2.2.0
ansible-core==2.12.6
ansible-lint==6.4.0
arrow==1.2.2
attrs==21.4.0
bcrypt==3.2.2
binaryornot==0.4.4
bracex==2.3
Cerberus==1.3.2
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
click-help-colors==0.9.1
colorama==0.4.4
commonmark==0.9.1
cookiecutter==2.1.1
cryptography==37.0.2
docker==5.0.3
enrich==1.2.7
flake8==4.0.1
future==0.18.2
gitchangelog==3.0.4
glob2==0.7
idna==3.3
importlib-resources==5.7.1
iniconfig==1.1.1
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.9.1
Mako==1.2.0
mando==0.6.4
MarkupSafe==2.1.1
mccabe==0.6.1
molecule==3.6.1
molecule-docker==1.1.0
packaging==21.3
parameterized==0.8.1
paramiko==2.11.0
parse==1.19.0
parse-type==0.6.0
pathspec==0.9.0
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
poyo==0.5.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.12.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
pystache==0.6.0
pytest==7.1.2
pytest-bdd==6.0.1
pytest-testinfra==6.7.0
python-dateutil==2.8.2
python-slugify==6.1.2
PyYAML==6.0
radon==5.1.0
requests==2.27.1
resolvelib==0.5.4
rich==12.4.4
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
subprocess-tee==0.3.5
tenacity==8.0.1
testinfra-bdd==2.0.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
wcmatch==8.3
websocket-client==1.3.2
yamllint==1.26.3
zipp==3.8.0
